requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
pydantic>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
PyJWT>=2.8.0
python-multipart>=0.0.6
onnx>=1.15.0
onnxruntime>=1.16.0
skl2onnx>=1.16.0
slowapi>=0.1.9

# Model backends
lightgbm>=4.0.0
scikit-learn>=1.3.0  # Fallback for testing
warpgbm  # GPU-accelerated gradient boosting (requires CUDA)

